Ignore CBMC proof failures, fail the build later
This commit ensures that the CBMC report still gets built even if the
proof fails. It makes Make ignore CBMC's return code, generating the
report unconditionally, but adds an extra check that reads CBMC's output
file to determine if there was a proof failure. This extra check fails
the entire pipeline if and only if a proof failed.