Move FreeRTOS_closesocket() out of vTCPStateChange
* Use vCloseSocket() instead of FreeRTOS_closesocket(),

* Rework the commit 86ee487a so that a state change
from eSYN_RECEIVED -> eCLOSE_WAIT makes the select
wake up and the user task can accept and close the
socket.

* Fix build with ipconfigTCP_HANG_PROTECTION == 0.

The problem is that a select will not wake up when the
child socket is accepted, but only when some data is received.
However if the server wants to send data immediately after
accpet that will not work.

* Also remove the code in prvTCPSocketCopy() which makes the
child socket inherit the socket set of the parent.
This is most likely just a work-around that makes the select
wake up when the first user data is received.