Add a check into BufferAllocation_2.c to ensure the addition of padding to the allocation size does not cause an integer overflow
Updated a broken link in README.md

In BufferAllocation_2.c
(https://github.com/FreeRTOS/FreeRTOS-Plus-TCP/blob/main/portable/BufferManagement/BufferAllocation_2.c),
there is an unchecked possible addition overflow when calculating the size of
the block of memory to be allocated for a network buffer that could result in
the size overflowing and the allocation returning success but allocating only
a fraction of the memory asked for. With default settings, this would only
occur when attempting to allocate within 12 bytes of 4 GB. Thanks to
Bernard Lebel (RMDS Innovation) for reporting this potential issue.

Signed-off-by: Aniruddha Kanhere <60444055+AniruddhaKanhere@users.noreply.github.com>