Add libslirp NetworkInterface implementation
Libslirp is used by qemu to provide userspace NAT networking. This serves as
a replacement for the current libpcap / winpcap NetworkInterface but does
not require elevated privileges or selecting a particular interface to bind to.

In this implementation, incoming connections are not supported.

Tested with git revision 38638382c8f3dc2523468f0c54014f4fdf0aa38f on MacOS and Windows.