Fix assertion failure when disconnecting the Ethernet cable (NXP1060 -  NetworkInterface.c) (#1002)
* Update NetworkInterface.c

The changes that I'm proposing are to make the driver to work in case of cable disconnection. 

In prvEMACHandlerTask, when xSendEventStructToIPTask, we have to pass network interface or it will fail the assertion in xSendEventStructToIPTask "configASSERT( pxInterface != NULL )".

In pxNXP1064_NetworkInterfaceInitialise, if xWaitPHY fails (for instance, the cable is not connected), the switch case will break without doing nothing and the prvEMACHandlerTask will be blocked waiting for the notification. We have to unblock it to try again in case of failure in xEMAC_WaitPHY.

* Update NetworkInterface.c

uncrustify code

* Update NetworkInterface.c

Correct indentation

* Update NetworkInterface.c

Correct indentation

* Update NetworkInterface.c

fix indentation... again

* Update source/portable/NetworkInterface/NXP1060/NetworkInterface.c

Co-authored-by: ActoryOu <jay2002824@gmail.com>

* Update source/portable/NetworkInterface/NXP1060/NetworkInterface.c

Co-authored-by: ActoryOu <jay2002824@gmail.com>

---------

Co-authored-by: ActoryOu <jay2002824@gmail.com>